requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
python-dotenv==1.0.0
PyMySQL==1.1.0
SQLAlchemy==2.0.21
numpy==1.24.3
transformers==4.37.2
tokenizers==0.15.1
torch==2.0.1
nltk==3.8.1
requests==2.31.0
gunicorn==21.2.0
Werkzeug==2.3.7
mysql-connector-python==8.2.0
PyPDF2==3.0.1
